added bulk delete

+++ resources/delete_live_servers.py
@@ -4,7 +4,7 @@ import shlex
 import csv
 import time
 
-
+ADMIN_START_IP=50
 START_IP = 60     #This is the start of the last octet that will be used for students in 192.168.10.0/24
 END_IP = 255
 class Student:
@@ -28,6 +28,9 @@ def get_vmid(NETID):
   return container_id
 
 def delete(container_id, NETID):
+  if container_id < 100:
+    print("The container ID must be at least 100. Please try again.")
+    exit()
   cmd = f"lxc-destroy -f {container_id}"
   res = subprocess.call(shlex.split(cmd), stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
   if res != 0 and res !=1:      #if there is an error
@@ -50,17 +53,18 @@ def delete(container_id, NETID):
       exit()
     print(f"Server deleted for {NETID}!")
 
-def delete_multiple(START_IP=START_IP, END_IP=END_IP):
-  RANGE_START = input("What IP do you want to start at?: 192.168.10.")
-  RANGE_START = input("What IP do you want to end at?: 192.168.10.")
-  confirm = input(f"Are you sure that you want to delete all servers between 192.168.10.{RANGE_START}-{RANGE_END}? (Y/N): ")
+def delete_multiple():
+  RANGE_START = input("What VM ID do you want to start at?:")
+  RANGE_END = input("What VM ID do you want to end at?:")
+  confirm = input(f"Are you sure that you want to delete all servers between {RANGE_START} and {RANGE_END}? (Y/N): ")
   if not confirm in ['Y', 'y']:
     print("Whew! Exiting...")
     exit()
-  for student in range(START_IP,END_IP):
-    
-    pass
-
+  to_delete = RANGE_START
+  for student in range(int(RANGE_START),int(RANGE_END)):
+    delete(int(to_delete), "VM ID:" + to_delete)
+    to_delete += 1
+  print(f"All servers between {RANGE_START} and {RANGE_END} were deleted!")
 
 def delete_one(NETID):
   #match student netID to container ID
